+++ envpool/mujoco/dmc/utils.cc
@@ -1,0 +1,111 @@
+// Copyright 2022 Garena Online Private Limited
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "envpool/mujoco/dmc/utils.h"
+
+#include <cmath>
+#include <fstream>
+#include <sstream>
+
+namespace mujoco_dmc {
+
+std::string GetFileContent(const std::string& base_path,
+                           const std::string& asset_name) {
+  // hardcode path here :(
+  std::string filename = base_path + "/mujoco/assets_dmc/" + asset_name;
+  std::ifstream ifs(filename);
+  std::stringstream ss;
+  ss << ifs.rdbuf();
+  return ss.str();
+}
+
+int GetQposId(mjModel* model, const std::string& name) {
+  return model->jnt_qposadr[mj_name2id(model, mjOBJ_JOINT, name.c_str())];
+}
+
+int GetSensorId(mjModel* model, const std::string& name) {
+  return model->sensor_adr[mj_name2id(model, mjOBJ_SENSOR, name.c_str())];
+}
+
+// rewards
+double RewardTolerance(double x, double bound_min, double bound_max,
+                       double margin, double value_at_margin,
+                       SigmoidType sigmoid_type) {
+  if (bound_min <= x && x <= bound_max) {
+    return 1.0;
+  }
+  if (margin <= 0.0) {
+    return 0.0;
+  }
+  x = (x < bound_min ? bound_min - x : x - bound_max) / margin;
+  if (sigmoid_type == SigmoidType::kGaussian) {
+    // scale = np.sqrt(-2 * np.log(value_at_1))
+    // return np.exp(-0.5 * (x*scale)**2)
+    double scaled_x = std::sqrt(-2 * std::log(value_at_margin)) * x;
+    return std::exp(-0.5 * scaled_x * scaled_x);
+  }
+  if (sigmoid_type == SigmoidType::kHyperbolic) {
+    // scale = np.arccosh(1/value_at_1)
+    // return 1 / np.cosh(x*scale)
+    double scaled_x = std::acosh(1 / value_at_margin) * x;
+    return 1 / std::cosh(scaled_x);
+  }
+  if (sigmoid_type == SigmoidType::kLongTail) {
+    // scale = np.sqrt(1/value_at_1 - 1)
+    // return 1 / ((x*scale)**2 + 1)
+    double scaled_x = std::sqrt(1 / value_at_margin - 1) * x;
+    return 1 / (scaled_x * scaled_x + 1);
+  }
+  if (sigmoid_type == SigmoidType::kReciprocal) {
+    // scale = 1/value_at_1 - 1
+    // return 1 / (abs(x)*scale + 1)
+    double scale = 1 / value_at_margin - 1;
+    return 1 / (std::abs(x) * scale + 1);
+  }
+  if (sigmoid_type == SigmoidType::kCosine) {
+    // scale = np.arccos(2*value_at_1 - 1) / np.pi
+    // scaled_x = x*scale
+    // with warnings.catch_warnings():
+    //   warnings.filterwarnings(
+    //       action='ignore', message='invalid value encountered in cos')
+    //   cos_pi_scaled_x = np.cos(np.pi*scaled_x)
+    // return np.where(abs(scaled_x) < 1, (1 + cos_pi_scaled_x)/2, 0.0)
+    const double pi = std::acos(-1);
+    double scaled_x = std::acos(2 * value_at_margin - 1) / pi * x;
+    return std::abs(scaled_x) < 1 ? (1 + std::cos(pi * scaled_x)) / 2 : 0.0;
+  }
+  if (sigmoid_type == SigmoidType::kLinear) {
+    // scale = 1-value_at_1
+    // scaled_x = x*scale
+    // return np.where(abs(scaled_x) < 1, 1 - scaled_x, 0.0)
+    double scaled_x = (1 - value_at_margin) * x;
+    return std::abs(scaled_x) < 1 ? 1 - scaled_x : 0.0;
+  }
+  if (sigmoid_type == SigmoidType::kQuadratic) {
+    // scale = np.sqrt(1-value_at_1)
+    // scaled_x = x*scale
+    // return np.where(abs(scaled_x) < 1, 1 - scaled_x**2, 0.0)
+    double scaled_x = std::sqrt(1 - value_at_margin) * x;
+    return std::abs(scaled_x) < 1 ? 1 - scaled_x * scaled_x : 0.0;
+  }
+  if (sigmoid_type == SigmoidType::kTanhSquared) {
+    // scale = np.arctanh(np.sqrt(1-value_at_1))
+    // return 1 - np.tanh(x*scale)**2
+    double scaled_x = std::atanh(std::sqrt(1 - value_at_margin)) * x;
+    return 1 - std::tanh(scaled_x) * std::tanh(scaled_x);
+  }
+  throw std::runtime_error("Unknown sigmoid type for RewardTolerance.");
+}
+
+}  // namespace mujoco_dmc
